+++ src/inlay.ts
@@ -45,13 +45,10 @@ const stateEffect = StateEffect.define<InlayHintEffect>(),
 					if (str === text) {
 						return inlayHints
 							? Decoration.set(
-								inlayHints.reverse()
-									.map(({position, label}) => [posToIndex(doc, position), label as string] as const)
-									.sort(([a], [b]) => a - b)
-									.map(
-										([index, label]) =>
-											Decoration.widget({widget: new InlayHintWidget(label)}).range(index),
-									),
+								inlayHints.map(({position, label}) => Decoration.widget({
+									widget: new InlayHintWidget(label as string),
+								}).range(posToIndex(doc, position))),
+								true,
 							)
 							: Decoration.none;
 					}