+++ src/lua.ts
@@ -1,7 +1,15 @@
 /* eslint-disable no-template-curly-in-string */
 import {lua} from '@codemirror/legacy-modes/mode/lua';
-import {syntaxTree, LanguageSupport, StreamLanguage, foldService} from '@codemirror/language';
+import {
+	syntaxTree,
+	LanguageSupport,
+	StreamLanguage,
+	foldService,
+	HighlightStyle,
+	syntaxHighlighting,
+} from '@codemirror/language';
 import {snippetCompletion} from '@codemirror/autocomplete';
+import {tags} from '@lezer/highlight';
 import {leadingSpaces, sliceDoc} from './util.js';
 import type {Extension} from '@codemirror/state';
 import type {CompletionSource, Completion} from '@codemirror/autocomplete';
@@ -431,6 +439,9 @@ const source: CompletionSource = context => {
 	return null;
 };
 const support: Extension = [
+	syntaxHighlighting(
+		HighlightStyle.define([{tag: tags.standard(tags.variableName), class: 'cm-globals'}]),
+	),
 	lang.data.of({autocomplete: source}),
 	foldService.of(({doc, tabSize}, start, from) => {
 		const {text, number} = doc.lineAt(start);